Health, GameControl, GUI

+++ Assets/Scripts/pmarin/GameControl.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 
 public static class GameControl {
-	public static int live = 100;
+	public static float live = 100f;
 	public static bool dead = false;
 	public static int score = 0;
 	//public static bool inmortal = false;
